Fixes for tests, added some new tests

+++ tests/RESTable.Tests/OperationsTests/ResourceAsync.cs
@@ -39,7 +39,7 @@ namespace RESTable.Tests.OperationsTests
             yield break;
         }
 
-        public async ValueTask<int> DeleteAsync(IRequest<ResourceAsync> request, CancellationToken cancellationToken)
+        public async ValueTask<long> DeleteAsync(IRequest<ResourceAsync> request, CancellationToken cancellationToken)
         {
             request.GetRequiredService<OperationsTestsFlags>().AsyncDeleterWasCalled = true;
             return 0;
